KIM-209 ♻️ 용도에 따라 클래스를 분리하다
