+++ src/main/java/com/devcourse/be04daangnmarket/member/api/MemberRestController.java
@@ -1,7 +1,6 @@
 package com.devcourse.be04daangnmarket.member.api;
 
 import com.devcourse.be04daangnmarket.common.auth.User;
-import com.devcourse.be04daangnmarket.common.jwt.JwtTokenProvider;
 import com.devcourse.be04daangnmarket.member.application.MemberService;
 import com.devcourse.be04daangnmarket.member.dto.MemberDto;
 import jakarta.validation.Valid;
@@ -11,13 +10,10 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -31,34 +27,10 @@ public class MemberRestController {
 
     private final MemberService memberService;
     private final PostService postService;
-    private final JwtTokenProvider jwtTokenProvider;
 
-    public MemberRestController(MemberService memberService, PostService postService, JwtTokenProvider jwtTokenProvider) {
+    public MemberRestController(MemberService memberService, PostService postService) {
         this.memberService = memberService;
         this.postService = postService;
-        this.jwtTokenProvider = jwtTokenProvider;
-    }
-
-    @PostMapping
-    public ResponseEntity<MemberDto.Response> signUp(@RequestBody @Valid MemberDto.SignUpRequest request) {
-        MemberDto.Response response = memberService.signUp(request);
-
-        return ResponseEntity.status(HttpStatus.CREATED)
-                .body(response);
-    }
-
-    @PostMapping("/login")
-    public ResponseEntity<MemberDto.Response> signIn(@RequestBody @Valid MemberDto.SignInRequest request) {
-        MemberDto.Response response = memberService.signIn(request);
-
-        String token = jwtTokenProvider.createToken(request.email());
-
-        HttpHeaders httpHeaders = new HttpHeaders();
-        httpHeaders.add("Authorization", token);
-
-        return ResponseEntity.status(HttpStatus.OK)
-                .headers(httpHeaders)
-                .body(response);
     }
 
     @GetMapping("/{id}")
