+++ src/main/java/com/devcourse/be04daangnmarket/member/view/MemberViewController.java
@@ -1,6 +1,5 @@
 package com.devcourse.be04daangnmarket.member.view;
 
-import com.devcourse.be04daangnmarket.member.application.KakaoService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -8,23 +7,6 @@ import org.springframework.web.bind.annotation.PathVariable;
 
 @Controller
 public class MemberViewController {
-    private final KakaoService kakaoService;
-
-    public MemberViewController(KakaoService kakaoService) {
-        this.kakaoService = kakaoService;
-    }
-
-    @GetMapping(value = "/sign-up")
-    public String signUp() {
-        return "signUp";
-    }
-
-    @GetMapping(value = "/sign-in")
-    public String signIn(Model model) {
-        model.addAttribute("kakaoUrl", kakaoService.getKakaoLogin());
-
-        return "signIn";
-    }
 
     @GetMapping(value = "/members/{id}/mypage")
     public String mypage(@PathVariable Long id, Model model) {
